requirements-devel.txt
requests>=2.32.2,<3.0.0
validators>=0.21.2,<1.0.0
authlib>=1.3.1,<2.0.0
grpcio>=1.57.0,<2.0.0
grpcio-tools>=1.57.0,<2.0.0

build
twine
wheel
importlib_metadata==7.2.1
setuptools_scm
Sphinx>=7.0.0
sphinx-rtd-theme==2.0.0
autodoc-pydantic==2.2.0
importlib_metadata==7.2.1

pytest==8.2.2
pytest-cov==5.0.0
pytest-asyncio==0.23.7
pytest-benchmark==4.0.0
pytest-profiling==1.7.0
coverage==7.6.0
pytest-xdist==3.6.1
werkzeug==3.0.3
pytest-httpserver==1.0.12
py-spy==0.3.14

numpy>=1.24.4,<2.0.0
pandas>=2.0.3,<3.0.0
pandas-stubs>=2.0.3,<3.0.0
polars>=0.20.26,<0.21.0

fastapi==0.111.1
flask[async]==3.0.3
litestar==2.9.1

mypy==1.11.0
mypy-extensions==1.0.0
tomli==2.0.1
types-protobuf==4.24.0.1
types-requests==2.30.0.0
types-urllib3==1.26.25.14
typing_extensions==4.8.0

pre-commit

flake8
flake8-bugbear==24.4.26
flake8-comprehensions==3.15.0
flake8-builtins==2.5.0
black==23.9.1